requirements.txt
absl-py==0.14.1
aiohttp==3.7.4.post0
appdirs==1.4.4
argon2-cffi==21.1.0
asgiref==3.5.0
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autoprop==4.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
bleach==4.1.0
body-measurements==1.1.1
Brotli==1.0.9
cachetools==4.2.4
calvis==0.0.1
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
chumpy==0.70
clang==5.0
click==8.0.3
colorlog==6.6.0
constantly==15.1.0
cryptography==35.0.0
cycler==0.10.0
Cython==0.29.26
dash==2.0.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.1
entrypoints==0.3
et-xmlfile==1.1.0
Flask==2.0.2
Flask-Compress==1.10.1
flatbuffers==1.12
gast==0.3.3
glooey==0.3.6
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==2.10.0
hbm==0.0.1
hyperlink==21.0.0
idna==3.2
image-classifiers==1.0.0
imageio==2.9.0
incremental==21.3.0
ipycanvas==0.9.1
ipyevents==2.0.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
ipyvtklink==0.2.1
ipywidgets==7.6.5
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.2
joblib==1.0.1
jsonschema==4.0.1
jupyter==1.0.0
jupyter-client==7.0.6
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
lxml==4.6.3
mapbox-earcut==1.0.0
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
meshio==4.4.6
meshtool==0.3
mistune==0.8.4
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.3
multidict==5.1.0
nbclient==0.5.4
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.4.60
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.0
Panda3D==1.10.10
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
plotly==5.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.20
protobuf==3.18.1
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycollada==0.7.2
pycparser==2.20
pyglet==1.5.21
Pygments==2.10.0
pyparsing==2.4.7
pyrsistent==0.18.0
pyslurm==20.11.8.0
python-dateutil==2.8.2
python-fcl==0.6.1
pytz==2021.3
pyvista==0.32.1
PyWavefront==1.3.3
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.2
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
Rtree==0.9.7
scikit-image==0.19.0
scikit-learn==1.0
scipy==1.7.2
scooby==0.5.7
Send2Trash==1.8.0
Shapely==1.8.0
signature_dispatch==1.0.0
six==1.15.0
spektral==1.0.8
sqlparse==0.4.2
svg.path==4.1
sympy==1.9
tenacity==8.0.1
tensorboard==2.2.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-estimator==2.2.0
tensorflow-gpu==2.2.0
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
torch==1.9.0
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
triangle==20200424
trimesh==3.9.35
Twisted==21.7.0
txaio==21.2.1
typeguard==2.13.3
typing-extensions==3.7.4.3
urllib3==1.26.7
vecrec==0.3.0
vedo==2021.0.7
vtk==9.0.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.2
widgetsnbextension==3.5.1
wrapt==1.12.1
wslink==1.0.7
xatlas==0.0.6
xxhash==2.0.2
yarl==1.6.3
zope.interface==5.4.0
